requirements.txt
fastapi
uvicorn
psycopg2-binary
psycopg-pool
python-dotenv
psycopg
alembic
sqlalchemy